requirements.txt
torch>=2.5.0
torchaudio>=2.5.0
faster-whisper>=1.0.0
pynput>=1.7.6
PyAudio>=0.2.13
numpy>=1.24.0
scipy>=1.11.0
pyperclip>=1.8.2
pystray>=0.19.5
Pillow>=10.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
